+++ example/bug-fixed/tests-compile-commands-json/b.cc
@@ -1,0 +1,5 @@
+#include <iostream>
+
+int main() {
+    std::cout << "b.cc run successfully!" << std::endl;
+}