+++ example/bug-fixed/watch-subdirectories-file-changes/xmake.lua
@@ -1,0 +1,4 @@
+target("test", function()
+    set_kind("binary")
+    add_files("watch_test/dir1/dir2/test.cc")
+end)