+++ xmake/core/package/package.lua
@@ -1620,7 +1620,7 @@ function _instance:buildhash()
                 table.sort(toolchains)
                 str = str .. "_" .. table.concat(toolchains, "_")
             end
-            return hash.uuid4(str):gsub('-', ''):lower()
+            return hash.strhash128(str)
         end
         local function _get_installdir(...)
             local name = self:name():lower():gsub("::", "_")
@@ -2561,6 +2561,20 @@ function _instance:check_mxxsnippets(snippets, opt)
     return sandbox_module.import("lib.detect.check_mxxsnippets", {anonymous = true})(snippets, opt)
 end
 
+-- check the given fortran snippets?
+--
+-- @param snippets  the snippets
+-- @param opt       the argument options, e.g. {configs = {defines = ""}}
+--
+-- @return          true or false, errors
+--
+function _instance:check_fcsnippets(snippets, opt)
+    opt = opt or {}
+    opt.target = self
+    opt.configs = self:_generate_build_configs(opt.configs, {sourcekind = "fc"})
+    return sandbox_module.import("lib.detect.check_fcsnippets", {anonymous = true})(snippets, opt)
+end
+
 -- the current mode is belong to the given modes?
 function package._api_is_mode(interp, ...)
     return config.is_mode(...)
@@ -2637,7 +2651,7 @@ function package.targetplat()
     return plat
 end
 
--- get global target architecture of pacakge
+-- get global target architecture of package
 function package.targetarch()
     local arch = package._memcache():get("target_arch")
     if arch == nil then