+++ xmake/core/sandbox/modules/interpreter/hash.lua
@@ -1,0 +1,2 @@
+-- load module
+return require("sandbox/modules/hash")