+++ xmake/modules/utils/ci/packageskey.lua
@@ -54,7 +54,5 @@ function main(requires_raw)
     end
     table.sort(keys)
     keys = table.concat(keys, ",")
-    print(hash.uuid4(keys):gsub('-', ''):lower())
+    print(hash.strhash128(keys))
 end
-
-