+++ xmake/plugins/project/clang/compile_commands.lua
@@ -28,6 +28,7 @@ import("private.utils.batchcmds")
 import("private.utils.executable_path")
 import("private.utils.rule_groups")
 import("plugins.project.utils.target_cmds", {rootdir = os.programdir()})
+import("actions.test.main", {rootdir = os.programdir(), alias = "test_action"})
 
 -- escape path
 function _escape_path(p)
@@ -248,6 +249,14 @@ function _add_targets(jsonfile)
             _add_target(jsonfile, target)
         end
     end
+    -- https://github.com/xmake-io/xmake/issues/4750
+    for _, test in pairs(test_action.get_tests()) do
+        local target = test.target
+        if not target:is_phony() then
+            _add_target(jsonfile, target)
+        end
+    end
+    jsonfile:print("")
     jsonfile:print("]")
 end